As discussed in another PR, `v.find` performance deteriorates as the config map grows and more config sources come in. We can significantly improve the performance of `v.find` operations (and thus all `Get*` operations) with a simple cache strategy.

As you can see in the benchmark below, not using a cache needs 4060 ns/op for reading all keys of a large config file, while using a cache reduces that time to 171ns/op on average.

Ristretto was chosen because it is a very performant, thread-safe
cache that does not suffer form congestion or long GC pauses.

```
% go test -bench=. .
goos: darwin
goarch: amd64
pkg: github.com/spf13/viper
BenchmarkFind/cache=false-8       289495              4060 ns/op
BenchmarkFind/cache=true-8       6980976               171 ns/op
PASS
```