@sagikazarmark 2020-02-20
Wow, this is super cool. I've been meaning to try ristretto for some time, but didn't have a chance.

I'll take a look at the PR soon! Thanks @aeneasr