@sagikazarmark 2020-02-20
Sure!

How did you find the race? Because the race detector is enabled in Viper too:

https://github.com/spf13/viper/blob/master/Makefile#L45