@aeneasr 2020-02-20
With a simple `go test -race -v ./...`. Maybe it has to do something with the fork, I'll investigate. But it seems to be related to `cache.Clear` and `cache.Get` getting racy.