@aeneasr 2020-02-20
Ah, I know why this isn't caught here. Enabling `-race` doesn't do anything really unless there are actual tests with concurrency. We have a test for that in our fork as viper was (still is?) quite racy when it comes to set/get. You can find the test here: https://github.com/ory/viper/blob/get-cache-ory/viper_test.go#L2340

I think viper will have a race condition if that test was added without this patch as well, as there are concurrent read/writes to the map.