@sagikazarmark 2020-02-20
Well, for now this is what it is, but for the long term I really want to separate reading and writing in Viper.

IMHO fetching config from various sources is what's Viper's strong suite, writing/setting config is just there, but not really the primary use case for Viper (or at least I think it shouldn't be).

So for the long term I plan to keep the current API for compatibility reasons, at least for one more major version, but after that I'd like to see writing either gone from the core, or moved a different interface entirely. Writing and Get-ing are usually done by totally different consumers, so it should be okay.