@aeneasr 2020-02-20
> IMHO fetching config from various sources is what's Viper's strong suite, writing/setting config is just there, but not really the primary use case for Viper (or at least I think it shouldn't be).

All of this works well until there's actually reloading (e.g. on config file changed) involved, which is where viper just breaks all over the place. We're using a RWMutex to prevent stuff like that but it obviously takes away some of the benefits you get from e.g. ristretto, although RLocks are kinda fine I guess.

> setting config is just there, but not really the primary use case for Viper (or at least I think it shouldn't be).

As long as setting values for testing is supported I agree :)