@sagikazarmark 2020-02-20
> All of this works well until there's actually reloading

Yeah, Viper pretty much sucks at that at the moment. I'm looking at changing that in a next version.

> As long as setting values for testing is supported I agree :)

Obviously. Manually setting values should be possible, but that's a different consumer of Viper.

Ideally "getters" should only be able to use the getter API and not set values. That would probably get rid of lots of races.